# Optimized Tokenizer.php