--- a/.travis.yml
+++ b/.travis.yml
@@ -1,18 +1,22 @@
 language: php
 
-# Setting sudo access to false will let Travis CI use containers rather than
-# VMs to run the tests. For more details see:
-# - http://docs.travis-ci.com/user/workers/container-based-infrastructure/
-# - http://docs.travis-ci.com/user/workers/standard-infrastructure/
 sudo: false
 
 php:
-  - 5.3
   - 5.4
   - 5.5
   - 5.6
   - 7.0
+  - 7.1
   - hhvm
+
+matrix:
+    include:
+        - php: 5.3
+          dist: precise
+    fast_finish: true
+
+dist: trusty
 
 cache:
   directories: