I am working on a project where we parse several thousand HTML5 pages with this library.
When profiling the code with xdebug, I noticed extreme amounts of calls to the scanner's .current() and .next() functions. Here is my attempt to optimize the tokenizer code.

Run time comparison:
Original Code: 281s
Only with the optimized quotedAttributeValue() function: 242s
With all optimizations:  226s 

-> Saves about ~60s / 20%