@goetas 2017-07-26
Hi, thanks for your PR. 

Looks good.. give me some time to check it properly