@mundschenk-at 2017-08-25
Any timeline on this? I know things like this can be quite intricate, but I'm about to release a new major version of wp-Typography and any speed increase in the tokenizer would be very, very welcome.