@goetas 2017-08-25
Most probably will have a look at it on Monday

On 25 Aug 2017 11:35, "Der Mundschenk & Compagnie" <notifications@github.com>
wrote:

> Any timeline on this? I know things like this can be quite intricate, but
> I'm about to release a new major version of wp-Typography and any speed
> increase in the tokenizer would be very, very welcome.
>
> —
> You are receiving this because you commented.
> Reply to this email directly, view it on GitHub
> <https://github.com/Masterminds/html5-php/pull/130#issuecomment-324870494>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AAvaJ1a8MBrmgIM8KZ7vugcFTR4iw2CAks5sbpV1gaJpZM4Ogewd>
> .
>