@goetas 2017-09-01
closed by https://github.com/Masterminds/html5-php/pull/135